fixed problem with coordinates of the physical bodies, they now spawn at the right position and the parameters in the physBodyFactory is all in pixels and the convertion is made in those methods

+++ dat255_group3/src/com/dat255_group3/utils/CoordinateConverter.java
@@ -18,7 +18,7 @@ public class CoordinateConverter {
 	}
 	
 	public static Vector2  meterToPixel(Vector2 meters){
-		float meterToPixel=CoordinateConverter.ScreenWidth/10;
+		float meterToPixel=(float)CoordinateConverter.ScreenWidth/10f;
 		Vector2 pixels = new Vector2 ();
 		pixels.x = meters.x*meterToPixel;
 		pixels.y = meters.y*meterToPixel;
